Reorganize standalone header location & process
Migrate split files to src/ then make include only contains
the standalone result.
