+++ src/spy.hpp
@@ -5,7 +5,8 @@
   SPDX-License-Identifier: MIT
 **/
 //==================================================================================================
-#pragma once
+#ifndef SPY_SPY_HPP_INCLUDED
+#define SPY_SPY_HPP_INCLUDED
 
 #include <spy/arch.hpp>
 #include <spy/compiler.hpp>
@@ -15,3 +16,5 @@
 #include <spy/sanitizers.hpp>
 #include <spy/simd.hpp>
 #include <spy/stdlib.hpp>
+
+#endif
