+++ src/spy/detail.hpp
