+++ src/spy/simd/ppc.hpp
